+++ ceammc/CHANGELOG.md
@@ -18,6 +18,7 @@
 - random.discrete added
 - array.set added
 - array.bpm added (via SoundTouch library: https://www.surina.net/soundtouch/)
+- array.stretch added (time-stretch, pitch-shift and rate-range) 
 
 ### Changed:
 - random.gauss doc update